Update size with bigint
It fixes #3 \n Signed-off-by: Diego Andrés <diegoandres_cortes@outlook.com>

+++ install.php
@@ -673,7 +673,7 @@
 			'{db_prefix}members', 
 			array(
 				'name' => 'shopBank',
-				'type' => 'int',
+				'type' => 'bigint',
 				'default' => 0,
 			)
 		);
